# stsiv-api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "media_ml",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/version/list": {
            "get": {
                "summary": "Get Experiment Versions",
                "description": "_summary_\n\nArgs:\n    current_user (CurrentUser, optional): _description_. Defaults to Depends(get_auth_user).\n    version_service (VersionService, optional): _description_. Defaults to Depends( BaseContainer().get_version_service ).\n\nReturns:\n    List[GetVersionSchema]: _description_",
                "operationId": "get_experiment_versions_api_v1_version_list_get",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ListVersionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/GetVersionSchema"
                                    },
                                    "type": "array",
                                    "title": "Response Get Experiment Versions Api V1 Version List Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/version/create": {
            "post": {
                "summary": "Create Experiment",
                "description": "_summary_\n\nArgs:\n    experiment_data (CreateExperimentRequest): _description_\n    current_user (CurrentUser, optional): _description_. Defaults to Depends(get_auth_user).\n\nReturns:\n    GetExperimentSchema: _description_",
                "operationId": "create_experiment_api_v1_version_create_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateVersionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetVersionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/version/{version_id}/upload": {
            "post": {
                "summary": "Upload File For Version",
                "description": "_summary_\n\nArgs:\n    version_id (UUID): _description_\n    file (UploadFile, optional): _description_. Defaults to File(...).\n    current_user (CurrentUser, optional): _description_. Defaults to Depends( get_auth_user).\n    is_data_file (bool, optional): _description_. Defaults to True.\n    version_service (VersionService, optional): _description_. \n    Defaults to Depends( BaseContainer().get_version_service).\n\nReturns:\n    _type_: _description_",
                "operationId": "upload_file_for_version_api_v1_version__version_id__upload_post",
                "parameters": [
                    {
                        "name": "version_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Version Id"
                        }
                    },
                    {
                        "name": "is_data_file",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Is Data File"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_file_for_version_api_v1_version__version_id__upload_post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/version/{version_id}/params": {
            "put": {
                "summary": "Update Versions Params",
                "description": "_summary_\n\nArgs:\n    version_id (UUID): _description_\n    params (VersionParamsRequest): _description_\n\nReturns:\n    GetVersionSchema: _description_",
                "operationId": "update_versions_params_api_v1_version__version_id__params_put",
                "parameters": [
                    {
                        "name": "version_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Version Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VersionParamsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetVersionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/version/{version_id}/run": {
            "post": {
                "summary": "Run Tests",
                "description": "_summary_\n\nArgs:\n    version_id (UUID): _description_\n    current_user (CurrentUser, optional): _description_. Defaults to Depends( get_auth_user).\n    version_service (VersionService, optional): _description_. Defaults to Depends( BaseContainer().get_version_service).\n\nReturns:\n    str: _description_",
                "operationId": "run_tests_api_v1_version__version_id__run_post",
                "parameters": [
                    {
                        "name": "version_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Version Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_upload_file_for_version_api_v1_version__version_id__upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_file_for_version_api_v1_version__version_id__upload_post"
            },
            "Command": {
                "properties": {
                    "alias": {
                        "type": "string",
                        "title": "Alias"
                    },
                    "value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "string"
                            },
                            {
                                "items": {
                                    "type": "integer"
                                },
                                "type": "array"
                            }
                        ],
                        "title": "Value"
                    }
                },
                "type": "object",
                "required": [
                    "alias",
                    "value"
                ],
                "title": "Command"
            },
...
            "CreateVersionRequest": {
                "properties": {
                    "experiment_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Experiment Id"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "In this version we changed..."
                    }
                },
                "type": "object",
                "title": "CreateVersionRequest",
                "description": "_summary_\n\nArgs:\n    BaseModel (_type_): _description_"
            },
...
            "GetVersionSchema": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "experiment_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Experiment Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "params": {
                        "title": "Params",
                        "description": "params for certain version"
                    }
                },
                "type": "object",
                "required": [
                    "params"
                ],
                "title": "GetVersionSchema",
                "description": "_summary_\n\nArgs:\n    BaseModel (_type_): _description_"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ListVersionRequest": {
                "properties": {
                    "experiment_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Experiment Id"
                    }
                },
                "type": "object",
                "title": "ListVersionRequest",
                "description": "_summary_\n\nArgs:\n    BaseModel (_type_): _description_"
            },
...
            "ParameterItem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "number"
                            }
                        ],
                        "title": "Value"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "value"
                ],
                "title": "ParameterItem"
            },
            "Parameters": {
                "properties": {
                    "alias": {
                        "type": "string",
                        "title": "Alias"
                    },
                    "blockFrequencyTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "nonOverlappingTemplateTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "overlappingTemplateTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "approximateEntropyTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "serialTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "linearComplexityTestBlockLength": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "numberOfBitcountRuns": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "uniformityBins": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "bitsToProcessPerIteration": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "uniformityCutoffLevel": {
                        "$ref": "#/components/schemas/ParameterItem"
                    },
                    "alphaConfidenceLevel": {
                        "$ref": "#/components/schemas/ParameterItem"
                    }
                },
                "type": "object",
                "required": [
                    "alias",
                    "blockFrequencyTestBlockLength",
                    "nonOverlappingTemplateTestBlockLength",
                    "overlappingTemplateTestBlockLength",
                    "approximateEntropyTestBlockLength",
                    "serialTestBlockLength",
                    "linearComplexityTestBlockLength",
                    "numberOfBitcountRuns",
                    "uniformityBins",
                    "bitsToProcessPerIteration",
                    "uniformityCutoffLevel",
                    "alphaConfidenceLevel"
                ],
                "title": "Parameters"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "VersionParamsRequest": {
                "properties": {
                    "tests": {
                        "$ref": "#/components/schemas/Command"
                    },
                    "parameters": {
                        "$ref": "#/components/schemas/Parameters"
                    },
                    "iterations": {
                        "$ref": "#/components/schemas/Command"
                    },
                    "createResultFiles": {
                        "$ref": "#/components/schemas/Command"
                    },
                    "bitcount": {
                        "$ref": "#/components/schemas/Command"
                    },
                    "numOfThreads": {
                        "$ref": "#/components/schemas/Command"
                    }
                },
                "type": "object",
                "required": [
                    "tests",
                    "parameters",
                    "iterations",
                    "createResultFiles",
                    "bitcount",
                    "numOfThreads"
                ],
                "title": "VersionParamsRequest"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}